Fix template build skip condition
Template's "release status" have different concept of a version - it
reports for the latest built version, as there is no real tag associated
with template build. So, when doing new build, do not check if the
template is "not released" because that's still about old version, not
the one that is going to be built. Instead, tighten build timestamp
condition to refuse building the same version twice.