Disarm sqv --not-after verification
The signature is made just before posting it, a slight clock mismatch
(even 1s) makes sqv reject the signature. We don't really care when the
signature was made, we care about the content (which also includes the
timestamp), so provide some large enough margin (+24h). Unfortunately,
there seems to be no option to disable this check completely in sqv.